Moved ReportFormat tests into new structure
