+++ tests/protocols/gmpv208/entities/report_formats/test_get_report_formats.py
@@ -17,7 +17,7 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 
-class GmpGetReportFormatsTestCase:
+class GmpGetReportFormatsTestMixin:
     def test_get_report_formats(self):
         self.gmp.get_report_formats()
