+++ tests/protocols/gmpv208/entities/report_formats/test_verify_report_format.py
@@ -23,33 +23,27 @@ from gvm.protocols.gmpv208.entities.report_formats import (
 )
 
 
-class GmpDeleteReportFormatTestCase:
-    def test_delete(self):
-        self.gmp.delete_report_format('a1')
+class GmpVerifyReportFormatTestMixin:
+    def test_verify(self):
+        self.gmp.verify_report_format('a1')
 
         self.connection.send.has_been_called_with(
-            '<delete_report_format report_format_id="a1" ultimate="0"/>'
-        )
-
-    def test_delete_ultimate(self):
-        self.gmp.delete_report_format('a1', ultimate=True)
-
-        self.connection.send.has_been_called_with(
-            '<delete_report_format report_format_id="a1" ultimate="1"/>'
+            '<verify_report_format report_format_id="a1"/>'
         )
 
     def test_missing_id(self):
         with self.assertRaises(GvmError):
-            self.gmp.delete_report_format(None)
+            self.gmp.verify_report_format(None)
 
         with self.assertRaises(GvmError):
-            self.gmp.delete_report_format('')
+            self.gmp.verify_report_format('')
 
-    def test_delete_with_type(self):
-        self.gmp.delete_report_format(ReportFormatType.SVG)
+    def test_verify_with_type(self):
+        self.gmp.verify_report_format(ReportFormatType.SVG)
 
         report_format_id = get_report_format_id_from_string('svg').value
         self.connection.send.has_been_called_with(
-            '<delete_report_format '
-            'report_format_id="{}" ultimate="0"/>'.format(report_format_id)
+            '<verify_report_format report_format_id="{}"/>'.format(
+                report_format_id
+            )
         )
